使用对数平滑公式计算风险值,减少高值对结果的影响

+++ src/main/java/com/example/daptech/entity/vo/PhoneUsVo.java
@@ -18,5 +18,5 @@ public class PhoneUsVo {
     private String location;
     private String information;  //需要提取详细信息,算法待定
 
-    private Integer value; //号码风险值
+    private Double value; //号码风险值
 }
